finder_requirements.txt
gradio>=4.28.0
PyPDF2>=3.0.1
langchain-core>=0.1.34
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=10.2.0
langchain-google-vertexai>=0.0.5
langchain-google-genai>=0.0.7
langchain-openai>=0.1.6
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.2
openai>=1.30.0
python-dotenv>=1.0.1
